:meat_on_bone: this has to be put on hold, gulp 4.0 is required. So I'm commiting b/c I want to come back to this, but its a whole ordeal to move to the newer version of gulp and I dont have time for that rabbit hole
